Not good to drop the security
Hey,
Because this is education for (perhaps) new users/admins, it is not recommended to teach them in a first lab how to bypass the security. I do not expect that follows Microsoft's values either. So when you set the execute policies, please use at least "RemoteSigned". The best is to use "All Signed", but then you should teach also how to sign the .ps1 scripts.