DefaultClassification should be defined
Hey,
When defining classifications, but not defining what would be the default set. This is very bad as, if you do not do so, then the order on "ClassificationList" defines what is the default classification. And this is something you do not inform about this unexpected behavior for readers.

So it should be mandatory - on education at least - to define classifications correctly without unexpected results.